Added files via upload
This is what I have done so far with our project. Some parts are there only for testing (mainly the .txt file since I don't know how to work databases) but it all works so far. Some things will be changed, like the log in page being replaced with the one Jean made.